Reconcile int-ish data types with formatting strings [#823]
New checks in clang-18/19 brought new findings, some of them reasonable :)

* "%x" style formatting requires an (unsigned int) variable
* numeric literals and macros are (int) by default
* results of math with (uint16_t), done in some cases, are upscaled into
  (int) by default
* char, unsigned or not, seems to be also upscaled into (int)

Overall, had to change formatting strings in some cases, variable types
in others (e.g. flags or notification types do not make sense as signed)
and added casting in a few places that remained.

+++ common/common.c
@@ -4298,7 +4298,7 @@ int match_regex_hex(const regex_t *preg, const int n)
 {
 	char	buf[10];
 
-	snprintf(buf, sizeof(buf), "%04x", n);
+	snprintf(buf, sizeof(buf), "%04x", (unsigned int)n);
 
 	return match_regex(preg, buf);
 }
